requirements.txt
torch>=2.2
numpy>=1.24
pyyaml>=6.0
tqdm>=4.66
deepspeed>=0.13
